Add support for previously incorrectly created volumes with the wrong FS
Previously we didn't respect the users fs choice since there was no mechanism
to transmit that choice to the share-manager. Starting with v1.2 we respect
the users chosen fs but previously incorrectly created volumes no longer
work since they will now try to mount via the users chosen fs. So in that
case where we detect a missmatch of the users fs and the existing fs we
use the existing fsType.

Longhorn #2991

Signed-off-by: Joshua Moody <joshua.moody@suse.com>
Signed-off-by: WeiZheChang <weizhe.chang@gmail.com>